[net config] Added net config module
 - OCF and DGRAM can now use the same network/BLE config API's
