+++ src/Makefile.base
@@ -42,6 +42,10 @@ export LDFLAGS += -u _printf_float
 ccflags-y += -DZJS_PRINT_FLOATS
 endif
 
+ifneq ($(BLE_ADDR), none)
+ccflags-y += -DZJS_CONFIG_BLE_ADDRESS=\"$(BLE_ADDR)\"
+endif
+
 obj-y += main.o \
          zjs_callbacks.o \
          zjs_common.o \
@@ -64,9 +68,9 @@ obj-$(ZJS_PERFORMANCE) += zjs_performance.o
 obj-$(ZJS_UART) += zjs_uart.o
 obj-$(ZJS_OCF) += zjs_ocf_client.o \
                   zjs_ocf_server.o \
-                  zjs_ocf_common.o \
-                  zjs_ocf_ble.o
+                  zjs_ocf_common.o
 obj-$(ZJS_TEST_PROMISE) += zjs_test_promise.o
+obj-$(ZJS_NET_CONFIG) += zjs_net_config.o
 
 # skip for now for frdm_k64f
 ifeq ($(BOARD), arduino_101)
