+++ src/main.c
@@ -24,9 +24,6 @@
 #include "zjs_timers.h"
 #ifdef BUILD_MODULE_OCF
 #include "zjs_ocf_common.h"
-#ifdef CONFIG_NET_L2_BLUETOOTH
-#include "zjs_ocf_ble.h"
-#endif
 #endif
 #ifdef BUILD_MODULE_BLE
 #include "zjs_ble.h"
@@ -175,9 +172,6 @@ int main(int argc, char *argv[])
 
 #ifdef BUILD_MODULE_OCF
     zjs_register_service_routine(NULL, main_poll_routine);
-#ifdef CONFIG_NET_L2_BLUETOOTH
-    zjs_init_ocf_ble();
-#endif
 #endif
 
 #ifndef ZJS_SNAPSHOT_BUILD
